Move DestroyDir from testutil to file_util
